requirements.txt
Django==4.2.1
djangorestframework==3.14.0
python-dotenv==1.0.0
drf-yasg==1.21.5
psycopg2-binary==2.9.6
djangorestframework-camel-case==1.4.2
pytest-django==4.5.2
django-allauth==0.52.0
dj-rest-auth==4.0.0
flake8==6.0.0

